+++ tests/test_store.py
@@ -14,6 +14,12 @@ class SerializationTestCase(TestCase):
             '{"hello": {"foo": "bar"}}',
         )
 
+    def test_serialize_logs_on_failure(self):
+        self.assertEqual(
+            store.serialize({"hello": {"foo": b"bar"}}),
+            '{"hello": {"foo": "bar"}}',
+        )
+
     def test_deserialize(self):
         self.assertEqual(
             store.deserialize('{"hello": {"foo": "bar"}}'),