+++ docs/changes.rst
@@ -5,6 +5,7 @@ Pending
 -------
 
 * Added a note about the default password in ``make example``.
+* Removed logging about the toolbar failing to serialize a value into JSON.
 
 6.0.0 (2025-07-22)
 ------------------